+++ src/evaluate/utils.rs
@@ -11,7 +11,7 @@ fn get_buf_reader(file_name: String) -> BufReader<File> {
     BufReader::new(file)
 }
 
-pub fn get_input_data(test_number: u32) -> (usize, Vec<f32>) {
+pub fn get_input_data(test_number: u32) -> (usize, Vec<f64>) {
     let file_name = format!("./input/input{:03}.txt", test_number);
     let reader = get_buf_reader(file_name);
 
@@ -19,9 +19,9 @@ pub fn get_input_data(test_number: u32) -> (usize, Vec<f32>) {
     let first_line = lines.next().expect("Empty file").unwrap();
     let n: usize = first_line.trim().parse().expect("Invalid number");
 
-    let mut radiuses: Vec<f32> = Vec::with_capacity(n);
+    let mut radiuses: Vec<f64> = Vec::with_capacity(n);
     for line in lines {
-        let radius: f32 = line
+        let radius: f64 = line
             .expect("Failed to read line")
             .trim()
             .parse()
@@ -33,7 +33,7 @@ pub fn get_input_data(test_number: u32) -> (usize, Vec<f32>) {
     return (n, radiuses);
 }
 
-pub fn get_jury_answer(test_number: u32) -> f32 {
+pub fn get_jury_answer(test_number: u32) -> f64 {
     let file_name = format!("./output/out{:03}.txt", test_number);
     let reader = get_buf_reader(file_name);
 
@@ -44,7 +44,7 @@ pub fn get_jury_answer(test_number: u32) -> f32 {
         .unwrap()
         .trim_end()
         .to_string()
-        .parse::<f32>()
+        .parse::<f64>()
         .expect("Valid float jury answer.")
 }
 
@@ -52,10 +52,10 @@ pub fn write_row_block(
     worksheet: &Arc<Mutex<&mut Worksheet>>,
     row: u32,
     col: u16,
-    main_circle_radius: f32,
+    main_circle_radius: f64,
     is_valid: bool,
-    points: f32,
-    time: f32,
+    points: f64,
+    time: f64,
     format: &Format,
 ) {
     worksheet
@@ -80,6 +80,6 @@ pub fn write_row_block(
         .ok();
 }
 
-pub fn calculate_points(answer: f32, jury_answer: f32) -> f32 {
+pub fn calculate_points(answer: f64, jury_answer: f64) -> f64 {
     ((2.0 - (answer / jury_answer)) * 100.0).max(0.0)
 }
