+++ src/ralgo/ralgo.rs
@@ -3,16 +3,16 @@ use nalgebra::{self};
 use super::calcfg::calcfg;
 
 pub fn ralg5(
-    mut x: nalgebra::DVector<f32>,
-    alpha: f32,
-    mut h: f32,
-    q1: f32,
-    epsx: f32,
-    epsg: f32,
+    mut x: nalgebra::DVector<f64>,
+    alpha: f64,
+    mut h: f64,
+    q1: f64,
+    epsx: f64,
+    epsg: f64,
     max_iterations: usize,
-    radiuses: &nalgebra::DVector<f32>,
-) -> nalgebra::DVector<f32> {
-    let mut b_matrix = nalgebra::DMatrix::<f32>::identity(x.len(), x.len());
+    radiuses: &nalgebra::DVector<f64>,
+) -> nalgebra::DVector<f64> {
+    let mut b_matrix = nalgebra::DMatrix::<f64>::identity(x.len(), x.len());
 
     let mut result_x = x.clone();
     let (mut result_f, mut g0) = calcfg(&result_x, radiuses);
@@ -22,13 +22,13 @@ pub fn ralg5(
     }
 
     for _ in 0..max_iterations {
-        let mut g1: nalgebra::DVector<f32> = b_matrix.tr_mul(&g0);
+        let mut g1: nalgebra::DVector<f64> = b_matrix.tr_mul(&g0);
 
-        let dx: nalgebra::DVector<f32> = (&b_matrix * &g1) / g1.norm();
+        let dx: nalgebra::DVector<f64> = (&b_matrix * &g1) / g1.norm();
         let dx_norm = dx.norm();
 
         let mut f;
-        let (mut d, mut ls, mut ddx) = (1.0_f32, 0_u32, 0.0_f32);
+        let (mut d, mut ls, mut ddx) = (1.0_f64, 0_u32, 0.0_f64);
         while d > 0.0 {
             x -= h * &dx;
             ddx += h * dx_norm;
@@ -73,16 +73,16 @@ pub fn ralg5(
 }
 
 pub fn ralgo_result_with_iterations(
-    mut x: nalgebra::DVector<f32>,
-    alpha: f32,
-    mut h: f32,
-    q1: f32,
-    epsx: f32,
-    epsg: f32,
+    mut x: nalgebra::DVector<f64>,
+    alpha: f64,
+    mut h: f64,
+    q1: f64,
+    epsx: f64,
+    epsg: f64,
     max_iterations: usize,
-    radiuses: &nalgebra::DVector<f32>,
-) -> (u32, u32, nalgebra::DVector<f32>) {
-    let mut b_matrix = nalgebra::DMatrix::<f32>::identity(x.len(), x.len());
+    radiuses: &nalgebra::DVector<f64>,
+) -> (u32, u32, nalgebra::DVector<f64>) {
+    let mut b_matrix = nalgebra::DMatrix::<f64>::identity(x.len(), x.len());
 
     let mut result_x = x.clone();
     let (mut result_f, mut g0) = calcfg(&result_x, radiuses);
@@ -93,13 +93,13 @@ pub fn ralgo_result_with_iterations(
     }
 
     for iter in 0..max_iterations as u32 {
-        let mut g1: nalgebra::DVector<f32> = b_matrix.tr_mul(&g0);
+        let mut g1: nalgebra::DVector<f64> = b_matrix.tr_mul(&g0);
 
-        let dx: nalgebra::DVector<f32> = (&b_matrix * &g1) / g1.norm();
+        let dx: nalgebra::DVector<f64> = (&b_matrix * &g1) / g1.norm();
         let dx_norm = dx.norm();
 
         let mut f;
-        let (mut d, mut ls, mut ddx) = (1.0_f32, 0_u32, 0.0_f32);
+        let (mut d, mut ls, mut ddx) = (1.0_f64, 0_u32, 0.0_f64);
         while d > 0.0 {
             x -= h * &dx;
             ddx += h * dx_norm;
